pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "daily-agent-garden"
version = "0.1.0"
description = "Reusable multi-agent framework for client projects using Google Vertex AI ADK"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "google-adk>=1.10.0",
    "google-genai>=1.9.0",
    "google-cloud-aiplatform>=1.60.0",
    "vertexai>=1.60.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["agents*", "shared*"]

[tool.pytest.ini_options]
testpaths = ["tests", "agents"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
